requirements.txt
appdirs==1.4.4
certifi==2020.12.5
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
numpy==1.19.3
opencv-contrib-python-headless==4.4.0.46
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==2.0.0
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
yapf==0.30.0
